refactor: divide bots em orquestrador, quente e frio para restaurar simetria e corrige bug de folga negativa
